extracted some code from CppCheckInspection to CppCheckInspectionImpl / some cleanups
